add from_iter.rs files
