+++ vortex-vector/src/bool/vector.rs
@@ -6,8 +6,7 @@
 use vortex_buffer::BitBuffer;
 use vortex_mask::Mask;
 
-use super::BoolVectorMut;
-use crate::VectorOps;
+use crate::{BoolVectorMut, VectorOps};
 
 /// An immutable vector of boolean values.
 ///
