+++ vortex-vector/src/primitive/mod.rs
@@ -26,6 +26,8 @@ pub use vector::PrimitiveVector;
 mod vector_mut;
 pub use vector_mut::PrimitiveVectorMut;
 
+mod from_iter;
+
 mod macros;
 
 use vortex_dtype::NativePType;
